+++ src/dirname/dirname.rs
@@ -1,4 +1,5 @@
 #![crate_name = "dirname"]
+#![allow(unstable)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -15,7 +16,7 @@ use std::io::print;
 
 static VERSION: &'static str = "1.0.0";
 
-pub fn uumain(args: Vec<String>) -> int {
+pub fn uumain(args: Vec<String>) -> isize {
     let program = args[0].clone();
     let opts = [
         getopts::optflag("z", "zero", "separate output with NUL rather than newline"),
