+++ src/kill/kill.rs
@@ -1,6 +1,5 @@
 #![crate_name = "kill"]
-#![feature(macro_rules)]
-#![feature(phase)]
+#![allow(unstable)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -17,7 +16,7 @@ extern crate libc;
 extern crate collections;
 extern crate serialize;
 
-#[phase(plugin, link)] extern crate log;
+#[macro_use] extern crate log;
 
 use std::io::process::Process;
 
@@ -32,6 +31,7 @@ use getopts::{
 use signals::ALL_SIGNALS;
 
 #[path = "../common/util.rs"]
+#[macro_use]
 mod util;
 
 #[path = "../common/signals.rs"]
@@ -40,8 +40,8 @@ mod signals;
 static NAME: &'static str = "kill";
 static VERSION:  &'static str = "0.0.1";
 
-static EXIT_OK:  int = 0;
-static EXIT_ERR: int = 1;
+static EXIT_OK:  isize = 0;
+static EXIT_ERR: isize = 1;
 
 pub enum Mode {
     Kill,
@@ -53,7 +53,7 @@ pub enum Mode {
 
 impl Copy for Mode {}
 
-pub fn uumain(args: Vec<String>) -> int {
+pub fn uumain(args: Vec<String>) -> isize {
     let opts = [
         optflag("h", "help", "display this help and exit"),
         optflag("V", "version", "output version information and exit"),
@@ -185,7 +185,7 @@ fn help(progname: &str, usage: &str) {
     println!("{}", get_help_text(progname, usage));
 }
 
-fn kill(signalname: &str, pids: std::vec::Vec<String>) -> int {
+fn kill(signalname: &str, pids: std::vec::Vec<String>) -> isize {
     let mut status = 0;
     let optional_signal_value = signals::signal_by_name_or_value(signalname);
     let signal_value = match optional_signal_value {
@@ -195,7 +195,7 @@ fn kill(signalname: &str, pids: std::vec::Vec<String>) -> int {
     for pid in pids.iter() {
         match pid.as_slice().parse() {
             Some(x) => {
-                let result = Process::kill(x, signal_value as int);
+                let result = Process::kill(x, signal_value as isize);
                 match result {
                     Ok(_) => (),
                     Err(f) => {
