+++ src/sleep/sleep.rs
@@ -1,4 +1,5 @@
 #![crate_name = "sleep"]
+#![allow(unstable)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -9,16 +10,15 @@
  * file that was distributed with this source code.
  */
 
-#![feature(macro_rules)]
-
 extern crate getopts;
 extern crate libc;
 
 use std::f64;
 use std::io::{print, timer};
-use std::time::duration::{mod, Duration};
+use std::time::duration::{self, Duration};
 
 #[path = "../common/util.rs"]
+#[macro_use]
 mod util;
 
 #[path = "../common/time.rs"]
@@ -26,7 +26,7 @@ mod time;
 
 static NAME: &'static str = "sleep";
 
-pub fn uumain(args: Vec<String>) -> int {
+pub fn uumain(args: Vec<String>) -> isize {
     let program = args[0].clone();
 
     let opts = [
