+++ src/unlink/unlink.rs
@@ -1,4 +1,5 @@
 #![crate_name = "unlink"]
+#![allow(unstable)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -11,21 +12,20 @@
 
 /* last synced with: unlink (GNU coreutils) 8.21 */
 
-#![feature(macro_rules)]
-
 extern crate getopts;
 extern crate libc;
 
 use std::io;
-use std::io::fs::{mod, PathExtensions};
+use std::io::fs::{self, PathExtensions};
 use std::io::print;
 
 #[path = "../common/util.rs"]
+#[macro_use]
 mod util;
 
 static NAME: &'static str = "unlink";
 
-pub fn uumain(args: Vec<String>) -> int {
+pub fn uumain(args: Vec<String>) -> isize {
     let program = args[0].clone();
     let opts = [
         getopts::optflag("h", "help", "display this help and exit"),
