+++ src/yes/yes.rs
@@ -1,4 +1,5 @@
 #![crate_name = "yes"]
+#![allow(unstable)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -11,19 +12,18 @@
 
 /* last synced with: yes (GNU coreutils) 8.13 */
 
-#![feature(macro_rules)]
-
 extern crate getopts;
 extern crate libc;
 
 use std::io::print;
 
 #[path = "../common/util.rs"]
+#[macro_use]
 mod util;
 
 static NAME: &'static str = "yes";
 
-pub fn uumain(args: Vec<String>) -> int {
+pub fn uumain(args: Vec<String>) -> isize {
     let program = args[0].clone();
     let opts = [
         getopts::optflag("h", "help", "display this help and exit"),
